requirements.txt
numpy
moviepy
google-cloud-videointelligence